Move sflow parser into input
