+++ plugins/inputs/sflow/parser/scripts/protodb/generate-embedded-data.notgo
