+++ plugins/inputs/sflow/parser/scripts/protodb/generated-embedded-data-go.template
