+++ plugins/inputs/sflow/parser/sflow/parser_test.go
@@ -7,7 +7,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/influxdata/telegraf/plugins/parsers/network_flow/decoder"
+	"github.com/influxdata/telegraf/plugins/inputs/sflow/parser/decoder"
 )
 
 func max(x, y int) int {
